requirements.txt
cohere==5.15.0
python-dotenv==1.1.0
streamlit==1.45.1